End of cleanup. Going to bug test, then get back to features.
