+++ Rpg.Mobile.App/App.xaml.cs
