+++ Rpg.Mobile.App/Game/MainBattle/StateMachines/Phases/SelectUnit/SelectUnitPhase.cs
@@ -1,13 +1,9 @@
 ﻿using Rpg.Mobile.App.Game.MainBattle.Data;
 
-namespace Rpg.Mobile.App.Game.MainBattle.States.Phases.SelectUnit;
+namespace Rpg.Mobile.App.Game.MainBattle.StateMachines.Phases.SelectUnit;
 
 public class SelectUnitPhase : IBattlePhase
 {
-    private readonly BattleData _data;
-
-    public SelectUnitPhase(BattleData data) => _data = data;
-
     public void Enter() { }
 
     public void Execute(float deltaTime) { }
