+++ Rpg.Mobile.App/Game/UserInterface/TextIndicatorComponent.cs
@@ -1,6 +1,6 @@
 ﻿using Rpg.Mobile.GameSdk.Tweening;
 
-namespace Rpg.Mobile.App.Game.Common;
+namespace Rpg.Mobile.App.Game.UserInterface;
 
 public class TextIndicatorComponent : TextboxComponent
 {
