+++ Rpg.Mobile.App/Platforms/Windows/App.xaml
@@ -1,8 +1,7 @@
 ﻿<maui:MauiWinUIApplication
-    x:Class="Rpg.Mobile.App.WinUI.App"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:maui="using:Microsoft.Maui"
-    xmlns:local="using:Rpg.Mobile.App.WinUI">
-
+     x:Class="Rpg.Mobile.App.WinUI.App"
+     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+     xmlns:maui="using:Microsoft.Maui"
+     xmlns:local="using:Rpg.Mobile.App.WinUI">
 </maui:MauiWinUIApplication>
