+++ Rpg.Mobile.App/Platforms/Windows/App.xaml.cs
@@ -3,6 +3,7 @@
 // To learn more about WinUI, the WinUI project structure,
 // and more about our project templates, see: http://aka.ms/winui-project-info.
 
+// ReSharper disable once CheckNamespace
 namespace Rpg.Mobile.App.WinUI;
 /// <summary>
 /// Provides application-specific behavior to supplement the default Application class.
@@ -20,4 +21,3 @@ public partial class App : MauiWinUIApplication
 
     protected override MauiApp CreateMauiApp() => MauiProgram.CreateMauiApp();
 }
-
