+++ Rpg.Mobile.App/Rpg.Mobile.App.csproj
@@ -107,8 +107,4 @@
 	  </None>
 	</ItemGroup>
 
-	<ItemGroup>
-	  <Folder Include="Game\MainBattle\States\Phases\Combat\" />
-	</ItemGroup>
-
 </Project>
